uploader_app/requirements.txt
Flask==2.3.2
python-dotenv==0.21.0
gunicorn==20.1.0
redis==4.6.0
ulid-py==1.1.0
